requirements.txt
snowflake-snowpark-python==1.5.0
click==7.1.2
